requirements.txt
streamlit==0.66.0
matplotlib==3.1.1
pandas == 0.25.1
plotly-express==0.4.1
sklearn

